Refactor: Move SparkHiveMetastore into 'spark' module
