+++ examples/src/test/scala/com/gu/tableversions/examples/DatePartitionedTableLoaderSpec.scala
@@ -8,8 +8,7 @@ import cats.effect.IO
 import com.gu.tableversions.core.Partition.PartitionColumn
 import com.gu.tableversions.core.TableVersions.UserId
 import com.gu.tableversions.core._
-import com.gu.tableversions.metastore.SparkHiveMetastore
-import com.gu.tableversions.spark.SparkHiveSuite
+import com.gu.tableversions.spark.{SparkHiveMetastore, SparkHiveSuite}
 import org.scalatest.{FlatSpec, Matchers}
 
 class DatePartitionedTableLoaderSpec extends FlatSpec with Matchers with SparkHiveSuite {
